api/requirements.txt
fastapi
pydantic
uvicorn
google-generativeai
python-dotenv
circle-developer-controlled-wallets
upstash-redis